Divisione Della customercard in base ed estensioni, aggiustamento del codice per comportarsi di conseguenza
