+++ src/main/java/it/unicam/cs/ids/lp/client/card/CustomerCardMapper.java
@@ -2,10 +2,11 @@ package it.unicam.cs.ids.lp.client.card;
 
 import it.unicam.cs.ids.lp.activity.card.CardRepository;
 import it.unicam.cs.ids.lp.client.CustomerRepository;
+import it.unicam.cs.ids.lp.client.card.programs.CashbackCard;
+import it.unicam.cs.ids.lp.client.card.programs.MembershipCard;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.UUID;
 import java.util.function.Function;
 
 @Service
@@ -18,14 +19,17 @@ public class CustomerCardMapper implements Function<CustomerCardRequest, Custome
 
     @Override
     public CustomerCard apply(CustomerCardRequest customerCardRequest) {
-        CustomerCard customerCard = new CustomerCard();
+        CustomerCard customerCard;
+        switch (customerCardRequest.program()) {
+            case CASHBACK -> customerCard = new CashbackCard();
+            case MEMBERSHIP -> customerCard = new MembershipCard();
+            default -> throw new IllegalStateException("Invalid CardProgram");
+        }
         customerCard.setCustomer(customerRepository.findById(customerCardRequest.customerId()).orElseThrow());
         customerCard.setCard(cardRepository.findById(customerCardRequest.cardId()).orElseThrow());
-        customerCard.setProgram(customerCardRequest.program());
         customerCard.setFamily(customerCardRequest.family());
-        customerCard.setReferred(customerCardRequest.referredCode());
-        customerCard.setReferralCode(UUID.randomUUID().toString());
         return customerCard;
     }
 }
+
