+++ src/main/java/it/unicam/cs/ids/lp/client/card/programs/CashbackCard.java
@@ -1,0 +1,23 @@
+package it.unicam.cs.ids.lp.client.card.programs;
+
+import it.unicam.cs.ids.lp.client.card.CustomerCard;
+import jakarta.persistence.Entity;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Getter
+@Setter
+@ToString(callSuper = true)
+@NoArgsConstructor
+public class CashbackCard extends CustomerCard {
+
+    private String cashbackReferred;
+
+    private String cashbackReferralCode;
+
+    private int remainingCashback;
+}
+
